Significant changes to sqio [xref J3/34,35,37,38,39,40,42,43]
Provides for reading subsequences, windows, and info-only, in addition
to complete sequences. [See flow chart, J3/37]

Sequential access functions:
Read
ReadInfo
ReadWindow

Random access functions:
Fetch
FetchInfo
FetchSubseq

ReadWindow() is capable of reading reverse strand after it's read the
forward strand, for nucleic acid seqs. (esl_sq_ReverseComplement() is
new.)

<ESL_SQ> has fields for tracking 1..L coord systems: start, end, C, W,
L. sq->n is now the actual length of the sequence stored in the
<ESL_SQ>, whereas sq->L is the total length of the original source
sequence.

Coordinates and lengths are now int64_t, not int; will support >2 Gb
seqs. Printing int64_t uses "%" PRId64 "...", not "%d" for printf()
format style. Functions that retrieve an (int *) result now need to
retrieve an (int64_t *) result.

SSI format is changed (magic number in ssi files is now "ssi3"). SSI
now stores lengths and key numbers as 64 bit int64_t. Some 64-bit
(off_t) stuff simplified in the code. 

esl-seqstat now uses esl_sqio_ReadWindow(), so it can stat large
sequences without reading them into memory all at once.

esl-sfetch can now do subsequence fetches, either singly (-c) or from
a list in a file (-Cf). -n renames a sequence. 

sqio's internal structure is generalized [see figure, J3/40] around
seebuf() (provides validation and residue counting); loadbuf() (loads
next buffer or line from memory); loadmem() (loads from disk into
memory cache); addbuf() (parses residues from buffer into ESL_SQ);
skipbuf() (skips number of residues in buffer). These functions now
work for all six input modes, and all three formats (fasta, genbank,
embl). Digital sqio input now validates against sqfp's inmap; invalid
characters are now detected as errors (previously, sqio wasn't
checking against digital alphabet properly).

sqio can now autodetect file format and alphabet from stdin or
gzipp'ed files, because it has a small data cache implemented
(sqfp->mem).

esl_sqfile_GuessFileFormat() improved; it tries looking at suffix
(.fa, for example).

esl_sqio_Rewind() removed; use esl_sqfile_Position(sqfp, (off_t) 0)
instead.

ESL_FAIL uses snprintf(errbuf, eslERRBUFSIZE), so is now robust
against overflow.

ESL_DSQ codes now include EOL, EOD; sqio parsers use this.

sqio has several unit tests now, but still could use more.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@255 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ easel.c
@@ -280,7 +280,7 @@ esl_usage(FILE *fp, char *progname, char *usage)
  * Date:     SRE, Thu May 13 10:56:28 1999 [St. Louis]
  *
  * Purpose:  Dynamic allocation version of fgets(),
- *           capable of reading unlimited line lengths.
+ *           capable of reading almost unlimited line lengths.
  *
  * Args:     buf - ptr to a string (may be reallocated)
  *           n   - ptr to current allocated length of buf,
@@ -395,7 +395,7 @@ esl_fgets(char **buf, int *n, FILE *fp)
  * Throws:   <eslEMEM> on allocation failure.
  */
 int
-esl_strdup(const char *s, int n, char **ret_dup)
+esl_strdup(const char *s, int64_t n, char **ret_dup)
 {
   int   status;
   char *new = NULL;
@@ -453,11 +453,11 @@ esl_strdup(const char *s, int n, char **ret_dup)
  *           is unaffected.
  */
 int
-esl_strcat(char **dest, int ldest, const char *src, int lsrc)
+esl_strcat(char **dest, int64_t ldest, const char *src, int64_t lsrc)
 {
-  void *p;
-  int   status;
-  int   len1, len2;
+  void     *p;
+  int       status;
+  int64_t   len1, len2;
 
   if (ldest < 0) len1 = ((*dest == NULL) ? 0 : strlen(*dest));
   else           len1 = ldest;
@@ -626,7 +626,7 @@ esl_strcasecmp(const char *s1, const char *s2)
  * Xref:      from squid's StringChop().
  */
 int
-esl_strchop(char *s, int n)
+esl_strchop(char *s, int64_t n)
 {
   int i;
   if (s == NULL) return eslOK;
@@ -668,10 +668,10 @@ esl_strchop(char *s, int n)
  * Returns:   <eslOK> on success.
  */
 int
-esl_strdealign(char *s, const char *aseq, const char *gapchars, int *opt_rlen)
+esl_strdealign(char *s, const char *aseq, const char *gapchars, int64_t *opt_rlen)
 {
-  int n = 0;
-  int apos;
+  int64_t n = 0;
+  int64_t apos;
 
   if (s == NULL) return eslOK;
 
@@ -683,6 +683,10 @@ esl_strdealign(char *s, const char *aseq, const char *gapchars, int *opt_rlen)
   if (opt_rlen != NULL) *opt_rlen = n;
   return eslOK;
 }
+
+
+
+
 
 /*----------------- end, C library replacements  -------------------------*/
 
@@ -1069,7 +1073,7 @@ esl_tmpfile(char *basename6X, FILE **ret_fp)
 /* Function:  esl_tmpfile_named()
  * Incept:    SRE, Sat Nov 11 09:13:25 2006 [Janelia]
  *
- * Purpose:   Open a persistant temporary file relative to the current
+ * Purpose:   Open a persistent temporary file relative to the current
  *            working directory. The file name is constructed from the
  *            <basename6X> argument, which must be a modifiable string
  *            ending in the six characters "XXXXXX".  These are
